+++ examples/example-general-renderer/src/main.rs
@@ -33,9 +33,8 @@
 //! ```
 
 use mingling::{
-    Groupped,
+    ChainProcess, Groupped,
     macros::{chain, dispatcher, gen_program, r_println, renderer},
-    marker::NextProcess,
     parser::Picker,
     setup::GeneralRendererSetup,
 };
@@ -61,7 +60,7 @@ struct Info {
 }
 
 #[chain]
-fn parse_render(prev: RenderCommandEntry) -> NextProcess {
+fn parse_render(prev: RenderCommandEntry) -> ChainProcess<ThisProgram> {
     let (name, age) = Picker::<()>::new(prev.inner)
         .pick::<String>(())
         .pick::<i32>(())